Add bcrypt and jsonwebtoken dependencies exercises all complete

+++ controllers/blogs.js
@@ -1,9 +1,11 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', (request, response) => {
     Blog
-      .find({})
+      .find({}).populate("user", { username: 1, name: 1 , id: 1})
       .then(blogs => {
         response.json(blogs)
       })
@@ -20,34 +22,44 @@ blogsRouter.get('/:id', (request, response, next) => {
 
   blogsRouter.post('/', async (request, response, next) => {
     const body = request.body;
-
-    if (!body.title || !body.url) {
-        return response.status(400).json({ error: 'title or url missing' });
-    }
-
-    const blog = new Blog({
+  
+    try {
+      const user = request.user;
+  
+      const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
         likes: body.likes || 0,
-    });
-
-    try {
-        const savedBlog = await blog.save();
-        response.status(201).json(savedBlog);
+        user: user._id
+      });
+  
+      const savedBlog = await blog.save();
+      user.blogs = user.blogs.concat(savedBlog._id);
+      await user.save();
+      response.status(201).json(savedBlog);
     } catch (error) {
-        next(error);
+      next(error);
     }
-});
+  });
 
   blogsRouter.delete('/:id', async (request, response, next) => {
     try {
-      const result = await Blog.findByIdAndDelete(request.params.id);
-      if (result) {
-        response.status(204).end();
-      } else {
-        response.status(404).end();
+      const user = request.user;
+  
+      const blog = await Blog.findById(request.params.id);
+      if (!blog) {
+        return response.status(404).json({ error: 'blog no encontrado' });
       }
+  
+      if (blog.user.toString() !== user._id.toString()) {
+        return response.status(403).json({ error: 'permiso denegado' });
+      }
+  
+      await Blog.findByIdAndDelete(request.params.id);
+      user.blogs = user.blogs.filter(blogId => blogId.toString() !== request.params.id);
+      await user.save();
+      response.status(204).end();
     } catch (error) {
       next(error);
     }
